Detect sleep
Add a systemd service which creates a file in /tmp just before the
system goes to sleep. When the system wakes up again, autosuspend checks
for the existence of this file in order to reset all timings. This
prevents that the system accidentally goes to sleep again directly after
waking up because all checks are long ago.