Switch versioning scheme
Provide a separate file to declare the version and the release of this
project which is used by setup.py and conf.py for the sphinx. This
ensures that both always use the same versions even if sphinx is called
without the setup.py command (e.g. on readthedocs).

This commit also switches the versioning scheme to semantic versioning
and prepares for the planned next major release of 1.0.0

fixes #14