Implement ical parsing logic
Stick together icalendar and dateutil's rrule to parse ical files for
events in an interesting time frame. This will be used to check for
events that currently block suspend and for events that will wake up the
server.

This was a severe nightmare to implement with all the little quirks
between dateutil and pytz (used by icalendar). I suspect this will break
with some version bump in the future but hopefully the test suite will
make fixing things easier.

This is the base for implementing #24.