Fix if clause in KodiIdleTime check (#54)
Fix if clause in KodiIdleTime check

Kodi sets the Idle boolean to ```true``` when the given time in ```System.IdleTime(x)``` runs out. Hence the user is active when the returned boolean is ```false```.
I've updated the if clause according to this observation.

(cherry picked from commit ba40d46af32cec4d62937563064ad30366283080)