Avoid includes in documentation
For being able to correctly link to individual checks, the existing
include structure crafted for the man pages had to be removed such that
the check descriptions are only included once in the main documentation.
This avoids a duplicate reference error when using labels in the check
configuration documents.

Man pages have been excluded from the main documentation now and make
use of toctree includes instead. Command line options were moved to the
main HTML documentation, because this information was previously only
included in the man pages.