Add a declarative config system for checks
Adds a new module with types for a lightweight config system for checks.
This is intended to provide a declaration of options for checks at
runtime such that configuration errors can be detect more easily at
program start.

The new config system will replace the create method for checks, which
make it easier to really use the inheritance of checks with their
mixins.