fix(activity): detect ipv4 mapped ipv6 connections
In the ActiveConnection check, properly handle the case where a service
listening on an IPv4 address is targeted by an IPv6 connection that
uses the special IPv6 range for mapping towards IPv4 addresses. This is
realized by transforming all IPv4 addresses into the IPv6 space.
Hopefully, this catches all corner-cases.

Fixes: #116